% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add_function.R
\name{add}
\alias{add}
\alias{add2}
\alias{add3}
\alias{add_quantiles}
\title{Adding Numbers and Square}
\usage{
add(x, y)

add2(x, y)

add3(x, y)

add_quantiles(x, y)
}
\arguments{
\item{x}{numeric: the first parameter in my add and square functio}

\item{y}{numeric: the second parameter}
}
\description{
Adding Numbers and Square
}
