% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageName-package.R
\docType{package}
\name{PackageName-package}
\alias{PackageName}
\alias{PackageName-package}
\title{PackageName: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Johan Larsson \email{johanlarsson@outlook.com} (\href{https://orcid.org/0000-0002-4029-5945}{ORCID})

}
\keyword{internal}
